+++ applications/fishing-map/src/features/workspace/workspace.selectors.ts
@@ -3,6 +3,7 @@ import { WorkspaceState } from 'types'
 import { RootState } from 'store'
 import { PUBLIC_SUFIX } from 'data/config'
 
+export const selectLastVisitedWorkspace = (state: RootState) => state.workspace.lastVisited
 export const selectWorkspace = (state: RootState) => state.workspace.data
 export const selectWorkspaceStatus = (state: RootState) => state.workspace.status
 export const selectWorkspaceCustomStatus = (state: RootState) => state.workspace.customStatus