+++ packages/layer-composer/src/generators/types.ts
@@ -28,9 +28,9 @@ export interface GeneratorFeature {
 }
 
 export interface GlobalGeneratorConfig {
-  start: string
-  end: string
-  zoom: number
+  start?: string
+  end?: string
+  zoom?: number
   token?: string
 }
 
@@ -277,8 +277,6 @@ export interface HeatmapAnimatedGeneratorConfig extends GeneratorConfig {
   datasetsStart?: string
   datasetsEnd?: string
   interactive?: boolean
-  staticStart?: string
-  staticEnd?: string
   /**
    * Defines a fixed or a supported list of intervals in an Array format
    */