+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @globalfishingwatch/ui-components
 
+## 4.1.0
+
+### Minor Changes
+
+- make the SplitView toggle optional
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [1a3d0960]
+  - @globalfishingwatch/layer-composer@7.0.0
+  - @globalfishingwatch/react-hooks@7.0.0
+
 ## 3.1.6
 
 ### Patch Changes